requirements.txt
bandit
Flask==2.0.1
pycryptodome==3.11.0
